+++ src/admin/utils.js
@@ -1,7 +1,13 @@
 import { MessageEmbed } from "discord.js";
 
 import { PERSONALITY } from "../personality.js";
-import { getAdminLogs, removeAdminLogs, removeBirthday, removeIgnoredUser, removeAlavirien } from "../helpers/index.js";
+import {
+  getAdminLogs,
+  removeAdminLogs,
+  removeBirthday,
+  removeIgnoredUser,
+  removeAlavirien,
+} from "../helpers/index.js";
 
 // jsons import
 import { readFileSync } from "fs";
@@ -234,7 +240,7 @@ export const generalEmbed = async (
   const aLog = personality.auditLog;
 
   const channel = await getLogChannel(obj); //get logChannel
-  if (process.env.DEBUG === "no" && checkProdTestMode(channel)) return; //if in prod && modif in test server
+  if (process.env.DEBUG === "no" && isTestServer(channel)) return; //if in prod && modif in test server
 
   const objToSend = objType === "user" ? obj.user : obj; //handle user objects case
   const embed = setupEmbed(color, perso, objToSend, embedType); //setup embed
@@ -688,7 +694,12 @@ export const octagonalLog = async (object, user) => {
   finishEmbed(octaPerso, null, embed, logChannel);
 };
 
-export const checkProdTestMode = (logChannel) => {
+/**
+ * Check if is currently in test server
+ * @param {Object} logChannel LogChannel retrieved from currentServer
+ * @returns True if is test server
+ */
+export const isTestServer = (logChannel) => {
   const server = commons.find(({ name }) => name === "test");
   const channels = [server.logChannelId, server.logThreadId];